pyproject.toml
[project]
name = "research-multi-agent"
version = "0.1.0"
description = "Research Multi-Agent System with LangChain and MCP"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "langchain-mcp-adapters>=0.1.0",
    "mcp>=1.0.0",
    "arxiv>=1.4.0",
    "youtube_search>=2.1.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "mcp_servers", "utils"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0"
]
